Resolved HTTP 500 by replacing preferences by functions.
