+++ app/src/main/java/org/fossasia/openevent/general/utils/Utils.kt
@@ -19,7 +19,9 @@ import androidx.fragment.app.FragmentManager
 import androidx.navigation.NavOptions
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import org.fossasia.openevent.general.R
+import org.fossasia.openevent.general.event.EventUtils.getSimpleFormattedDate
 import timber.log.Timber
+import java.util.Calendar
 import java.util.regex.Pattern
 
 object Utils {
@@ -137,6 +139,56 @@ object Utils {
             navigation.visibility = View.GONE
             navigation.animation = AnimationUtils.loadAnimation(context, R.anim.slide_down)
         }
+    }
+
+    fun getNextDate(): String {
+        val calendar = Calendar.getInstance()
+        calendar.add(Calendar.DATE, 1)
+        return getSimpleFormattedDate(calendar.time)
+    }
+
+    fun getNextToNextDate(): String {
+        val calendar = Calendar.getInstance()
+        calendar.add(Calendar.DATE, 2)
+        return getSimpleFormattedDate(calendar.time)
+    }
+
+    fun getWeekendDate(): String {
+        val calendar = Calendar.getInstance()
+        val today = calendar.get(Calendar.DAY_OF_WEEK)
+        if (today != Calendar.SATURDAY) {
+            val offset = Calendar.SATURDAY - today
+            calendar.add(Calendar.DATE, offset)
+        }
+        return getSimpleFormattedDate(calendar.time)
+    }
+
+    fun getNextToWeekendDate(): String {
+        val calendar = Calendar.getInstance()
+        val today = calendar.get(Calendar.DAY_OF_WEEK)
+        if (today != Calendar.SATURDAY) {
+            val offset = Calendar.SATURDAY - today
+            calendar.add(Calendar.DATE, offset)
+        }
+        calendar.add(Calendar.DATE, 1)
+        return getSimpleFormattedDate(calendar.time)
+    }
+
+    fun getNextMonth(): String {
+        val calendar = Calendar.getInstance()
+        val today = calendar.get(Calendar.DAY_OF_MONTH)
+        val offset = 30 - today
+        calendar.add(Calendar.DATE, offset)
+        return getSimpleFormattedDate(calendar.time)
+    }
+
+    fun getNextToNextMonth(): String {
+        val calendar = Calendar.getInstance()
+        val today = calendar.get(Calendar.DAY_OF_MONTH)
+        val offset = 30 - today
+        calendar.add(Calendar.DATE, offset)
+        calendar.add(Calendar.MONTH, 1)
+        return getSimpleFormattedDate(calendar.time)
     }
 
     enum class cardType {
